+++ src/Model/Declaration.ts
@@ -22,4 +22,8 @@ export default class Declaration {
 		this.weight = 1
 		this.normalizedWeight = 1
 	}
+
+	get normalizedPath() {
+		return JSON.stringify(this.path.filter(field => typeof field === "string"))
+	}
 }