+++ src/index.ts
@@ -2,7 +2,6 @@ import Jhaystack from "./Jhaystack"
 import * as ComparisonStrategy from "./Comparison/ComparisonStrategy"
 import * as ExtractionStrategy from "./Extraction/ExtractionStrategy"
 import * as SortingStrategy from "./Sorting/SortingStrategy"
-import * as IndexStrategy from "./indexing/IndexStrategy"
 import * as PreProcessingStrategy from "./PreProcessing/PreProcessingStrategy"
 
-export { Jhaystack, ComparisonStrategy, ExtractionStrategy, SortingStrategy, IndexStrategy, PreProcessingStrategy }
+export { Jhaystack, ComparisonStrategy, ExtractionStrategy, SortingStrategy, PreProcessingStrategy }